Create Multiple Interfaces implementation
When the objects of the Pet and Dog interfaces are created, individually, their functionalities overridden from the main reflect after the program is executed, by calling their methods.
Dog Interface has implemented the Pet interface, resulting in inheriting all the methods of the Pet interface.
Now, if I do not create the Object for the Pet interface, I can still access the Test() method of the Pet interface. This is done by using the object of the Dog interface. This is possible only because the Dog interface has implemented the method of Pet interface.

Output:
This is the Dog interface.
This is the Pet interface.